set up BlockList component and inital styling for sections
